Backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
easyocr
pytesseract
ultralytics
python-multipart